Updated compiler options

+++ liblpe/data/models/base/PluginRecallStrategies.cpp
@@ -38,7 +38,7 @@ PluginRecallStrategies::PluginRecallStrategies() {
     mDefaultStrategy = (PluginRecallStrategy) GetPrivateProfileInt(
             "general",
             "default",
-            (int) PluginRecallStrategy::PRESET,
+            PluginRecallStrategy::PRESET,
             resPath.data()
     );
